requirements.txt
langfuse==3.5.0
fastapi==0.117.1
uvicorn==0.37.0
openai==1.109.1
python-dotenv==1.1.1
pydantic==2.11.9
httpx==0.28.1